+++ templates/test/package.js.ts
@@ -1,5 +1,5 @@
 import { readFile, TemplateFunction } from "../../src/lib/createAdapter";
 
 export = (answers => {
-	return readFile("testStartup.raw.js", __dirname);
+	return readFile("package.raw.js", __dirname);
 }) as TemplateFunction;